+++ test/unit/sessions/video-session-handler.test.ts
@@ -1313,7 +1313,7 @@ describe('pinParticipantVideo', () => {
     await expect(handler.pinParticipantVideo(session, targetParticipcant)).rejects.toThrowError();
 
     expect(eventSpy).not.toHaveBeenCalled();
-    expect(errorSpy).toHaveBeenCalledWith(SdkErrorTypes.generic, 'Request to pin video failed', {
+    expect(errorSpy).toHaveBeenCalledWith(SdkErrorTypes.session, 'Request to pin video failed', {
       conversationId: session.conversationId,
       sessionId: session.id,
       error: fakeErr